Add method coupon() for payoff.
The coupon payoff represents a deterministic payoff of value at a given time.
This is of particular importance as terminal as Stack does not allow addition
of terminal values and there is no way to now the difference between a coupon
payment, or the optimal choice of possible terminal derivative payoffs.