Implement the Finite Difference Equation valuation method.
The FDEModel supports calculating the price of an option for a range of stock
values simultaneously, whereas the BinomialModel only supports calculations
for a single stock price.

Various iterative methods may be used by the FDEModel.  The current schemes
supported are the famous Crank-Nicolson method, and the standard Explicit and
Implicit methods.  The framework should allow for easy customisation or creation
of schemes used, or modification of the tridiagonal matrices used by the
schemes.

While here refactor the Value object to be shared between both BinomialModel
and FDEModel.