Add code to check prices using the FDE model.
Based on tests it is apparent that the FDE model is wrong.  The total default
prices are higher than the partial default.  This is inconsistent with the
pricing based on the binomial model.

The pricing without default (\lambda = 0) is largely consistent with the
binomial model but still not consistent with the AFV03 (table 3) figures.
Further investigation is required.