Refine convertible bond implementation to behave like AFV03.
The put (P) and call (C) provisions use "dirty" strike prices.  The outstanding
coupon payment, on a simple linear basis, is added to the "clean" strike price.
On coupon payment times it is assumed the coupon is payed before the options
are evaluated, thus no accrued coupons are needed.

The Annuity's (B) final coupon is payed as part of the nominal amount, not as
an explicit coupon.  This reduced the final payoff when the conversion option
takes affect.