Implement the R84 modification to Crank Nicolson to improve smoothness.
The Rannacher method uses four fully implicit quarter steps for the Crank-
NicolsonScheme whenever a signification discontinuity occurs, namely at the
terminal node and at each coupon payment.

The smoothing affect is particularly prevalent in the AFV03 figures 7 (gamma)
and less so in the AFV03 figure 6 delta.  This is definitely is a justifiable
improvement.