Add the Penalty-Rannacher scheme.
The PenaltyRannacherScheme uses four fully implicit quarter steps for the
terminal payment and for every coupon payment otherwise it uses the Crank-
Nicolson scheme with implicit American conditions.

The convergence is slightly slower than the PenaltyScheme but has stable prices
(as seen in the delta and gamma of the prices).